Add blueprint-modification logic for copy-paste operations
When player uses Ctrl-C "copy" tool to select an area, only `on_player_setup_blueprint` fires.
When `on_player_configured_blueprint` fires, it can be either when the area is first selected (and blueprint data is in `event.blueprint_to_setup`), or when the player accepts the blueprint dialog and is given the blueprint in the cursor.
This code adds handling of `on_player_setup_blueprint` for copy-paste operations, and handling of `event.blueprint_to_setup` so that items are replaced before the user sees the blueprint creation window.